Added tax calculation implementation service

+++ tax/src/main/java/com/dufuna/berlin/kutembachinambu/tax/exceptions/TaxBracketNotFoundException.java
@@ -1,7 +1,7 @@
 package com.dufuna.berlin.kutembachinambu.tax.exceptions;
 
 public class TaxBracketNotFoundException extends RuntimeException{
-    public TaxBracketNotFoundException(Long id) {
+    public TaxBracketNotFoundException(Integer id) {
         super("Could not find employee " + id);
     }
 }
